refactor(delete-user-data): restore legacy file layout
